requirements.txt
streamlit==1.12.2
pandas==1.4.4
snowflake-connector-python==2.8.0